pyproject.toml
[project]
name = "ecommerce-drf-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref==3.8.1",
    "astroid==3.3.10",
    "dill==0.4.0",
    "django==5.2.1",
    "django-cors-headers==4.7.0",
    "django-environ==0.12.0",
    "django-filter==25.1",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "gunicorn>=23.0.0",
    "isort==6.0.1",
    "markdown==3.8",
    "mccabe==0.7.0",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "pydantic[email]>=2.11.7",
    "pylint==3.3.7",
    "pylint-django==2.6.1",
    "pylint-plugin-utils==0.8.2",
    "sqlparse==0.5.3",
    "tomlkit==0.13.2",
]
